Fixed a bug that caused the read_assignments.csv to not include some cell barcode combinations.
The write in the loop was always printing the previously processed set of cell barcodes when it encountered a new/different cell barcode. At the end of the file, there is no new cell barcode combination, so the data for the last cell barcode combo is never printed to the output file. In most cases for large datasets or with poor quality sequencing, this is not a major issue, but with high quality sequencing and/or small datasets (e.g. test runs), large amounts of data can be omitted. In our case, 3 of 16 major barcode combinations were missing from the read_assignments file. This fix simply adds processing for the last cell barcode combination after the loop.

Files checked in: split_seq/processv2.py